Move from firebase to superstatic for serve
For serve_local.sh, `firebase` is an overkill, and it needs logged-in user. `superstatic` is Firebase’s CLI for serving that doesn’t require login.

This is more convenient when you have more than one Firebase accounts and would need to switch to the correct one anytime you want to test the site.

More importantly, it unlocks he possibility to test the site (`linkcheck`) on Travis, as is the case in the site-webdev repo.