Redo guidelines around getters and setters. (#728)
* Redo guidelines around getters and setters.

This is a lot of new prose. :( Deciding when to use a getter versus a
method seems to be difficult for many Dart users, so I felt it was
worth going into detail.

I also tweaked a bunch of examples and added a couple of other related
guidelines.

Fix #298. Fix #569.

* Revise.

* Fix analysis error in excerpt.

* Refresh excerpt.

* Simplify excerpt.