Change the recommendation for doc comments (#969)
DartDoc will show setter comments as a property. It is only when there is a comment on the setter and a getter that the comment is ambiguous, and the getter is favored. Change the recommendation so that teams can comment the setter if it makes more sense for the readability of the code.