Single quotes on raw string example in lang tour
This is a tiny nit, but I just ran into trouble trying to figure out what this statement from `built_value` was doing:

      StandardJsonPlugin({this.discriminator = r'$', this.valueKey = ''});

I'd never seen the leading-r syntax for a raw string before. I Cmd-F'ed the site for "r'" and came up empty because the example is using double quotes instead. If we generally encourage single quotes as the default (which is implied by the other examples), could we change this example to match?