Add `environment` field to sdk constraint section (#1082)
In the sdk constraint section of the Dart 2 page it previously showed only the `sdk` field, which must be nested under the `environment` field to work. It also didn't have the proper indentation so it looked like a top level key.

This adds the `environment` key with proper indentation for the `sdk` field, which should help clarify how to set up sdk constraints.