Always recommend using constraints in pubspec.yaml
I have seen issues in the ecosystem based on the latest versions of
multiple packages conflicting with each other, or the latest version of
a package not quite being compatible with another arbitrary version of
Flutter. I think even for apps it's better to recommend locking specific
known versions in `pubspec.yaml` even if a `pubspec.lock` committed to
source would make this redundant in the general case.