Document a CHANGELOG.md format for use on pub.dev
New upcoming design of pub.dev will "parse" the changelog. And display
the section headers for versions differently. In the future more tooling
might parse or read changelogs. So this is an attempt to document a
format for changelogs that can be easily parsed.

This does not provide a formal specification of all CHANGELOG.md formats
accepted by pub.dev, but it should provide sufficient documentation that
package authors can write CHANGELOG.md files that will be parsed
correctly on pub.dev.