Assert: throw error rather than exception
Given 

- the difference in Dart of Exception vs Error
- and `assert` throws an `AssertionError`

It seems more clear to say "throw an error" rather than "throw an exception" when used in the context of `assert`. 

However, I may be mistaken about the meaning of "throw an exception" if that phrase is broad enough to include errors. If so, though, that's not very clear in the documentation.