Use local (untested) snippets in dartpad_picker
Upgrading to Jekyll 3.9 caused the <?code-excerpt> tags to be
transformed into paragraph elements. Running the same Jekyll Converter
to process code excerpts fixed this issue. However, the snippets on
the homepage. By default the snippets in _try-dart-examples.md don't
get processed by this converter. Removing the underscore gets them
to be processed by this converter, but are added to the homepage as
prettified snippets, which isn't something dartpad_picker can use.

This is a temporary fix to allow us to upgrade to Jekyll 3.9:
https://github.com/dart-lang/site-www/issues/2578