Fix some null safety errors in "Effective Dart" example code context.
I wanted to get a sense of how many of the examples would be affected
by null safety, so I temporarily enabled that and went through the
errors. The good news is that it's not bad. Probably half a dozen with
trivial fixes.

The bad news is that those fixes do require using null safety feature
syntax in the examples themselves ("?", "required", etc.). So I'm not
going to make those changes yet.

But, while I was at it, I went ahead and fixed a bunch of null safety
errors in the surrounding context code. These changes are entirely
backwards compatible and fine to land now on master so I figured I'd
send them out first.