Remove mention of dartdevc from analysis options page (#3063)
I'm not sure this mention of a resource makes a lot of sense anymore as it's a tool for compiling and debugging web applications rather than a static analysis tool. 

I believe in the past this page had a lot of mentions of strong mode and how dartdevc only supported it or something along those line, but those mentions aren't relevant and do not exist anymore.