Updating null safety codelab example
The third exercise in the null safety codelab was assigning a value to `aNullableListOfStrings`, and as a result that variable wasn't functioning as it should in the exercise.

`aNullableListOfStrings` and `aListOfNullableStrings` are intended (at least back when I did the snippets) to teach people that both `List<String>?` and `List<String?>` are valid, but used for different things. If `aNullableListOfStrings` is assigned a value, you never need to change its type to `List<String>?` to make the analyzer happy.