Update information about web libraries (#3164)
WebSQL's standard was dropped 5+ years ago and is unsupported on most major browsers except Chrome. We'll be deprecating it too, but we feel there is no need to highlight it here.

`js_util` is not quite deprecated, it is still the recommended API to workaround limitations in `dart:html` and `package:js`, so we want to make the labeling here more accurate.