Fix errors in FAQ question about List.filled
There was a syntax error (`List?<Foo?>`).  Also the example talked about initializing a list with the same value when in fact the "bad" example was initializing each element of the list with a different value.  This change fixes the syntax error and splits the one example into two (one using `List.filled`, for when there really is the same value, and the other using `List.generate`).