Fix h1 overlapping subnav
Fixes it by allowing the subnav to take on its necessary height if its multiple lines long.

I recommend viewing the preview and verifying pages with the page subnav (back and forward sections at the top) are formatted as you expect. Then do the same by switching to a mobile layout with developer tools.

Fixes #3582