Update Default parameter values information
The current text implies that it is possible to define default values for both types of positional parameters (required and optional), as it says: "Your function can use = to define default values for both named and positional parameters.". However, positional parameters can be of 2 types, required and optional, and only the optional ones allow the definition of a default value.

This change is intended to make it clear that only named and optional positional parameters are likely to have default values set.