Move default value for optional parameters discussion to each individual section (#4366)
Currently setting default values for optional named and positional
parameters was separated, but it is pretty core to each topic, so this
PR moves the discussion of default values to each of the parameter
sections. This helps the examples make a bit more sense as they are
contextualized.

I've also removed the mention of migrating from `=` to `:` because it
was added over 6 years ago, most code already uses `:`, the extra
mention is large and potentially confusing (and don't want to duplicate
it in this new layout), and in Dart 2.19 the old syntax will become a
warning anyway.

Also removes the mention of how list and maps can be passed. It doesn't
talk about sets or other constant objects which can be passed as well
and isn't super relevant to the goals of these sections. We should
instead expand on our documentation of const
(https://github.com/dart-lang/site-www/issues/4350) elsewhere and link
to that.

Overall:
- Moves default value discussion to named and optional positional
sections instead of isolated
- Better clarifies the nullable types and when they are needed, as well
as how `null` is the default default value
- Moves discussion of named parameters anywhere to end
- Removes mention of old `=` syntax
- Removes mention of how list and maps can be passed as don't want to
duplicate it and it is better documented by const documentation
elsewhere

Fixes https://github.com/dart-lang/site-www/issues/3737

Co-authored-by: Anthony Sansone <atsansone@users.noreply.github.com>