Switch to opt-out system for code block language tags (#4469)
This works by adding an extra class to the body on pages in which
language tags should be included. Code blocks can add the `no-lang`
class with `{:no-lang}` to avoid the language tag being added.

Classes which want the language tags to show up need to add `body_class:
highlight-languages` to the front matter.