Use Dart Sass for Jekyll's sass conversions (#4488)
The old `sassc` gem uses libsass which is deprecated, replaced by Dart
sass which is used by sass-embedded and jekyll-sass-converter 3+.

This PR updates to use Dart Sass, to help keep new changes compatible
with future Sass and Jekyll versions but also as a step towards
https://github.com/dart-lang/site-www/issues/4164.

Using `/` as a division operator [is
deprecated](https://sass-lang.com/documentation/breaking-changes/slash-div),
so the options are `calc(x / y)`, `math.div(x, y)`, or in some cases it
makes sense to switch to multiplication instead. This PR chooses
switching to multiplication for most instances.