[2.19] New `pub add` yaml syntax (#4551)
Fixes #4514 

`add` has a new yaml syntax that mostly replaces the `--option`
arguments. You can still use the old syntax and arguments, which I tried
to maintain in this rewrite, but the better way is to use yaml (and the
old way will be eventually deprecated(?))

Honestly, this rewrite got a little bloated/messy and I'm not sure if I
like it. It was hard to preserve the old syntax and document the new
(though we decided [it's important to preserve the old
syntax](https://github.com/dart-lang/site-www/issues/4514#issuecomment-1399330752)).
The way I restructure it is:
- Basic `add` intro
  - Still works, so didn't change 
- Except elaborating on constraint options a little more, which might be
unnecessary and can be removed if we want.
- YAML section (new)
  - Introduce the new syntax
  - Sub sections for yaml sources
- Each source corresponds to an old --option, so at the end of each
section there's a line that says _Previously <--option>_
    - If the old option had an example, I preserved it
- In case any of the options' sections were linked to, I preserved the
section heading anchor with the new section names
- Options section (old)
- There are still options that work, that haven't been replaced by yaml,
so those are still there
- As mentioned, the replaced options info are now in the source sections

---------

Co-authored-by: Parker Lougheed <parlough@gmail.com>