Improve word breaking/wrapping in TOC on underscores (#4562)
Currently in the right-hand TOC (table of contents), on long strings
with underscores (`_`), such as on the linter rules and diagnostic
messages page, the entries are just wrapped/broken on the first
character necessary. This PR introduces a line break opportunity element
([`<wbr>`](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/wbr))
after each underscore (`_`) in the TOC. This allows the browser to
prefer to wrap just after an underscore if its needed. This results in
the following benefits:

- Words aren't broken in half, making the entries easier to read
- The underscore being at the end indicates to the reader that the entry
continues on the next line
- It looks nicer and more consistent

I've accomplished this with a new Liquid filter named
`underscore_breaker`, which can be used later for pages with underscores
in the title, such as if we move diagnostics to their own individual
pages.