Move 'Generators' to 'Functions' for now (#4653)
It's not currently linked to from the sidenav, so it isn't really
accessible.

Since we have all the function content in one place, there isn't really
a great place to put it in the sidenav. For now, we can put it at the
bottom of the functions page since it is less commonly used. What do you
think?