docs: Use absolute path for PUB_CACHE env variable
The description "home directory" can lead to confusion. It can be stated clearer where to find the cache directory by using the absolute path with `$HOME` environment variable. This also avoids issues with `relative path`, or miswriting the path e.g. with `~` when manually setting it in env variables (`.bashrc`, etc.)