Use absolute path for `PUB_CACHE` environment variable (#4752)
The description "home directory" can lead to confusion. It can be stated
clearer where to find the cache directory by using the absolute path
with `$HOME` environment variable. This also avoids issues with
`relative path`, or miswriting the path e.g. with `~` when manually
setting it in env variables (`.bashrc`, etc.)

Closes https://github.com/dart-lang/pub/issues/3861

---------

Co-authored-by: Parker Lougheed <parlough@gmail.com>