Suggest using getters in mixins
I've long understood how mixins could provide value, but it took me way to long to understand how to access state in the super class (which makes mixins far more powerful). I'm not sure if this rewording is enough or if it may be helpful to add a simple example of a getter or accessing a member of the superclass in the mixin.

I like the brief docs, but this is a case where one more example would've gone a long way for me.