The PUB_CACHE have moved.
This migration started in Dart 2.8.0, since then we've used the old
location of the PUB_CACHE if it already existed at that location.
But users who have started using Dart since 2.8.0 should be unaffected
as we started creating the PUB_CACHE in the new location, when no
PUB_CACHE was present anywhere.