Update typedefs.md, changed function typedef return type
The last function typedef example declares an int as the return type. This seemed silly to me since the inputs of the function are parameter types, implying it is intended to be a generic function typedef. 
For my edit, I've replaced the "int" with "T" since that appears to be have been the intent.