Replace usages of Font Awesome with Material Icons (#5013)
This replaces the remaining usages of Font Awesome with Material Icons
and removes Font Awesome. The way we are using Font Awesome is resulting
in a 436kB download then JS running on top of that. Before even
considering the JS parsing/running, it accounts for 318ms out of 545ms
of the site load on my internet on Chrome Dev 116.

Fixes https://github.com/dart-lang/site-www/issues/3790