Redirect away from unused articles index page (#5180)
The [articles index](https://dart.dev/articles) is a remnant of the past
when some docs and some blog posts were published as articles on the
site. It no longer serves a purpose as it's not in the sidenav and only
links to one article. This PR removes the page and redirects it the
Medium blog.

The remaining articles under `/articles` will be slowly removed or moved
elsewhere when they are updated.

Contributes to https://github.com/dart-lang/site-www/issues/5177