Remove old cheatsheet (#5241)
This [old cheatsheet](https://dart.dev/guides/language/cheatsheet) is
very minimal and isn't super useful as it stands, since it hasn't been
maintained beyond its initial inclusion. It's also not really linked to
from anywhere. There is value in having a quick reference, but this
isn't it.

This PR redirects the page to the [Language intro
page](https://dart.dev/language) instead of the [Cheatsheet
codelab](https://dart.dev/codelabs/dart-cheatsheet) because the samples
on the intro page serve a similar function as a (relatively) quick
reference with just text. I'm happy to consider other options though :)

**Old link to test redirect:**
https://dart-dev--pr5241-feature-remove-old-c-he3mwjob.web.app/guides/language/cheatsheet