[3.2] Understanding null safety page updates for private final field promotion (#5325)
---

Part of #4254 

---

Questions / notes: I'm not sure if the wording in these parts need to be
slightly adjusted, or are unrelated.

- Line 405: 
> **Top level variable and static field declarations must have an
initializer.** Since these can be accessed and assigned from anywhere in
the program, it’s impossible for the compiler to guarantee that the
variable has been given a value before it gets used. The only safe
option is to require the declaration itself to have an initializing
expression that produces a value of the right type:

- not sure if this statement is affected, because later is says that
local variables are the most flexible case, I'm assuming because they
can be type promoted? So are the rules about fields now also more
flexible, at least for private final fields, since they can be promoted
now too?

- [Late
variables](https://dart.dev/null-safety/understanding-null-safety#late-variables)
- "The most common place where the type checker cannot prove the safety
of code is around top-level variables and fields." ... technically still
true, maybe adjust?