Adjust error message text to match implementation.
https://dart-review.googlesource.com/c/sdk/+/337609 made a small
change in the context message that the analyzer and CFE generate when
field promotion fails due to the member in question being
public. Previously the message referred to the non-promoting members
as a `public field`. Now it refers to it as a `public property`
(because in certain circumstances, private properties *are*
promotable).

This small change in the context message text doesn't really have any
impact on the advice we want to give to users about field
promotability, but it's nice for the documentation to match the
implementation's output as closely as possible.