Update type-system.md
In the Runtime checks Section the description says that " the following code throws an exception at runtime because it's an error to cast a list of dogs to a list of cats", but in the given code the type is of List itself. I've edited that. That's it. A simple typo. ;)