Remove null safety codelab for now (#5539)
Removes the null safety codelab and references to it. In a few days, I
will follow up soon with some new samples on the cheatsheet to replace
this.

This codelab is the least viewed on the site and no longer makes sense
to separate from other language learning. As mentioned, I will follow up
with a few new samples on the cheatsheet as a stop-gap until we
implement a consolidated new-learner experience.

Contributes to https://github.com/dart-lang/site-www/issues/5382
Contributes to https://github.com/dart-lang/site-www/issues/4603
Closes https://github.com/dart-lang/site-www/issues/3093