Update type-system.md (#5549)
As mentioned by @eernstg in this
[comment](https://github.com/dart-lang/site-www/pull/5523#discussion_r1485628164),
I've updated the code which corrects the confusion caused here. In the
doc, static type of animals is List<Animal>, but the run-time type of
animals is a type that implements List<Dog>.

Fixes [Link to
page](https://dart.dev/language/type-system#runtime-checks)

---

- [ ] I’ve reviewed the contributor guide and applied the relevant
portions to this PR.
- [ ] This PR doesn't contain automatically generated corrections or
text (Grammarly, LLMs, and similar).
- [ ] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn't use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [ ] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.