Add note about if-null operator (#5649)
Wanted to address a side point @dtonhofer made in #5635 

> Additionally, the "if null" operator `a ?? b` is apparently generally
called a
["null-coalescing"](https://en.wikipedia.org/wiki/Null_coalescing_operator)
operator. One might want to note that.

Not trying to rename it, it seems like we use if-null across the board.
But I confirmed we do use "null coalescing" to describe our if-null
operator in the spec at least once:

<img width="461" alt="image"
src="https://github.com/dart-lang/site-www/assets/111139605/eb308b65-dea0-4d69-9eba-a0f55232a7b2">

So I think it makes sense to say "also known as the null coalescing
operator" in at least one place, for people more familiar with that
term.