Provide more details on error when Markdown code block is misformatted (#5743)
@MaryaBelanger I believe
https://github.com/dart-lang/site-www/issues/5732 is caused by a code
block without a language specified in your local branch in the
`/src/content/language/macros.md` file. This PR makes that more obvious
by providing an error explaining that.

I'll follow-up with a better error message, but wanted to make sure this
actually was the issue you're facing. I'm also open to changing the
requirement of a language if that's helpful or important to you. Thanks!