Remove outdated numeric computation article (#5776)
This page is out of date and has been marked as obsolete for a long
time. It was written when there were three representations and goes in
to too many internal details, many of which are no longer correct and
definitely not relevant for most Dart developers. The page also barely
gets any views.

There's perhaps some value in the performance suggestions discussed
here, but we should revisit that from a wholistic perspective that is
more cross-platform, particularly for today's VM and the Wasm target.

So this PR removes the article, redirecting to the closest other page we
have, that discusses numbers on different platforms.

Contributes to https://github.com/dart-lang/site-www/issues/5766 by
removing one of the few remaining archived articles.