Best practices for downgrade testing (#5838)
Testing with `dart pub downgrade` is an optional extra thing a package
author can do.

There is naturally a huge combination of different versions one can test
with. But testing with the latest (`dart pub upgrade`) and oldest (`dart
pub downgrade`) is a good choice.

**for package authors**!

For application developers it doesn't make much sense, it's better to
simply rely on `pubspec.lock` and perhaps liberally apply `dart pub
upgrade --tighten` as there is rarely any need to downgrade dependencies
(once they have been upgraded).

For **package authors** allowing a wide dependency constraint is more
important, otherwise it becomes too easy for package consumers to run
into a dependency conflict. Where constraints from different packages
they want to use are mutually incompatible.

Note: `dart pub upgrade` won't give the latest packages, it'll try, but
it'll still respect incompatible version constraints. The same is the
case for `dart pub downgrade`, it tried to give the oldest dependencies,
but it may just give an oldest set of dependencies. The version solver
does not search the entire state-space, it makes one decision at the
time, and when it reaches a solution it's happy. `upgrade/downgrade`
really only determines which branches of the solution space it tries to
explore first.

---------

Co-authored-by: Sigurd Meldgaard <sigurdm@google.com>
Co-authored-by: Parker Lougheed <parlough@gmail.com>
Co-authored-by: Marya <111139605+MaryaBelanger@users.noreply.github.com>