Account for removal of lint groups (#6033)
Lints no longer are part of three specific groups, but now have
tags/groups (https://github.com/dart-lang/linter/issues/1020). This PR
removes the expectation of the three groups so that the lints keep
rendering. It also updates the source data to the latest version with no
groups.

As part of the work in
https://github.com/dart-lang/site-www/issues/4499, we'll surface the new
categories/tags later on :)

Closes https://github.com/dart-lang/sdk/issues/56396