feat(docs): Add documentation for using records and typedef
This commit introduces a new section in the documentation showcasing the
combination of Dart's record types and `typedef` to create simple,
reusable data structure for dynamic UI rendering.

This approach was originally shared as an answer on Stack Overflow in
Spring 2023 to the question: "Is there anything like a struct in Dart?"
(https://stackoverflow.com/a/76391196/12158825). The solution received
highly positive feedback via emails and comments, highlighting its
usefulness for developers.

This feature is now documented as a recommended approach for developers
aiming to write cleaner and more maintainable Dart code that wasn't
mentioned before.