Explaining the concepts of Label (break and continue) in Dart  (#6451)
This PR improves the Dart documentation by adding detailed explanations
and examples for the ```break``` and ```continue``` statements in
various loop structures (for, while, do-while, and nested loops).

The changes have made in two directories:

1) ```/examples/language/lib/control_flow/branches.dart```

2) ```/src/content/language/branches.md```

Preview:


https://dart-dev--pr6451-j-manoj-06-patch-4-fywn6bb1.web.app/language/loops#labels

- [x] I’ve reviewed the contributor guide and applied the relevant
portions to this PR.
- [x] This PR doesn't contain automatically generated corrections or
text (Grammarly, LLMs, and similar).
- [x] This PR follows the [Google Developer Documentation Style
Guidelines](https://developers.google.com/style) — for example, it
doesn't use _i.e._ or _e.g._, and it avoids _I_ and _we_ (first person).
- [x] This PR uses [semantic line
breaks](https://github.com/dart-lang/site-shared/blob/main/doc/writing-for-dart-and-flutter-websites.md#semantic-line-breaks)
of 80 characters or fewer.

This ensures and follows the contributing guidelines.
As per the issue , the Labels in Dart is explained with examples.
Waiting for further improvements.

Fixes #6289

Best regards,
Manoj J

---------

Co-authored-by: Amanda Fitch <18406675+antfitch@users.noreply.github.com>
Co-authored-by: Parker Lougheed <parlough@gmail.com>