Documentation style: Document conceptual properties as properties (#6522)
Say explicitly that a function that is named like a property (noun
phrase or non-imperative verb phrase, as described in `design.md`)
should also be documented like a property (noun-phrase or "Whether ...",
basically).

(I keep seeing people documenting a noun-phrased function with "Returns
noun-phrase", just because the style guide does not have any exceptions
for using verb-phrases for functions. If it's treated as a property,
which is allowed, it should also be documented as a property.)

Tweaks some examples to not contradict other rules. For example, don't
have a two-statement first paragraph "good" example after the "single
statement first paragraph" rule.)

(Will update good/bad examples too..)

---------

Co-authored-by: Parker Lougheed <parlough@gmail.com>
Co-authored-by: Amanda Fitch <18406675+antfitch@users.noreply.github.com>