Break up Language > Syntax basics for improved learning order (#6564)
Reorganizes the **Language >** section of the sidenav by breaking up the
**Syntax basics** section. The category was a sort of catch all and
didn't all represent syntax. Beyond that, some of the contents didn't
make sense in that order, particularly **Libraries & imports** and
**Metadata** which make much more sense after having learned at least
functions. Then **Keywords** is a separate reference/index, so place it
at the end of the sidenav category.

Fixes https://github.com/dart-lang/site-www/issues/4654