Add a note to also migrate examples and docs (#6569)
These docs are mostly archived and legacy now due to the public
migration being largely complete, so I don't think
https://github.com/dart-lang/site-www/issues/2776 is a large issue. Just
in case, this PR adds a small note to also migrate examples and docs.

https://github.com/dart-lang/site-www/issues/2776 also mentions a case
of dependencies of examples not being updated for null safety, but I
don't think that's a likely case anymore that we should worry about.

Closes https://github.com/dart-lang/site-www/issues/2776

---------

Co-authored-by: Amanda Fitch <18406675+antfitch@users.noreply.github.com>