Individual pages for diagnostic docs (#6573)
Following the split we did for linter rule documentation, move all
diagnostic docs to their own page at
`/tools/diagnostics/<diagnostic_name>`. This allows for increased focus
and improved loading time when a developer navigates to the
documentation for a diagnostic from their IDE.

As the docs are moved to individual pages, also introduce a new index of
diagnostics at `/tools/diagnostics`. This shares some styles with the
[linter rule index](https://dart.dev/tools/linter-rules#rules) and now
includes diagnostics without extra documentation to serve as a more
complete index. Each diagnostic is included as a card with: its name,
its problem messages, a copy button, and a learn more button if it has a
full page.

Contributes to https://github.com/dart-lang/site-www/issues/4499

**Staged index:**
https://dart-dev--pr6573-feat-diagnostic-message-pages-kre5no9y.web.app/tools/diagnostics
**Staged non-lint diagnostic:**
https://dart-dev--pr6573-feat-diagnostic-message-pages-kre5no9y.web.app/tools/diagnostics/definitely_unassigned_late_local_variable
**Staged lint diagnostic:**
https://dart-dev--pr6573-feat-diagnostic-message-pages-kre5no9y.web.app/tools/diagnostics/prefer_final_locals

---------

Co-authored-by: Amanda Fitch <18406675+antfitch@users.noreply.github.com>